remove interpreter
