--- yarn-project/acir-simulator/src/avm/interpreter/index.ts
@@ -1,1 +1,0 @@
-export * from './interpreter.js';
